fix(ios): fix bug that caused ios app to crash when searching for notes
This commit fixes an issue that caused the ios app to crash when
searching for a note. The issue arised from the fact that
Dispatchers.Default was used to in HomeScreen ViewModel to get the filtered
items. While this works on Android, on IOS, using any dispatchers
other than Main on flows, caused the app to crash.
Eg:
someFlow.flowOn(Dispatchers.Default) // this will make the app crash on IOS. Only Dispatchers.Main will work.
	.launchIn(viewModelScope)

So, this commit makes the DispatcherProvider instance have all
properties set to Dispatchers.Main in the ios portion of the
shared module.